Feat: 프로필 페이지 구현

+++ src/components/templates/ProfileTemplate.tsx
@@ -1,19 +1,22 @@
-import { useMemo } from 'react';
+import { useMemo, useState } from 'react';
 import styled from 'styled-components';
-import Container from '../common/Container';
-import Text from '../common/Text';
-import Flex from '../common/Flex';
-import Avatar from '../common/Avatar';
-import Image from '../common/Image';
-import Button from '../common/Button';
 import {
   IconMessage,
   IconUserMinus,
   IconUserPlus,
   IconUsers,
 } from '@tabler/icons-react';
+import Container from '../common/Container';
+import Text from '../common/Text';
+import Flex from '../common/Flex';
+import Avatar from '../common/Avatar';
+import Image from '../common/Image';
+import Button from '../common/Button';
 import { UserType } from '~/types/common';
 import { OptionalConfig } from '~/hooks/api';
+import Modal from '../common/Modal';
+import ProfileImageUpload from '../uploadImage/ProfileImageUpload';
+import UserListItem from '../UserListItem';
 
 export interface ProfileTemplatePropsType {
   user: UserType;
@@ -22,10 +25,17 @@ export interface ProfileTemplatePropsType {
     requestUser: (config?: OptionalConfig) => Promise<void>;
     createFollow: (config?: OptionalConfig) => Promise<void>;
     deleteFollow: (config?: OptionalConfig) => Promise<void>;
+    uploadPhoto: (config?: OptionalConfig) => Promise<void>;
   };
 }
 
 const ProfileTemplate = ({ user, auth, actions }: ProfileTemplatePropsType) => {
+  const [isShowUpload, setIsShowUpload] = useState(false);
+  const [isLoadingUpload, setIsLoadingUpload] = useState(false);
+  const [isShowProfile, setIsShowProfile] = useState(false);
+  const [isShowFollower, setIsShowFollower] = useState(false);
+  const [isShowFollowing, setIsShowFollowing] = useState(false);
+
   const isMe = useMemo(() => {
     return user._id === auth?._id;
   }, [user, auth]);
@@ -35,6 +45,8 @@ const ProfileTemplate = ({ user, auth, actions }: ProfileTemplatePropsType) => {
   }, [user, auth]);
 
   const handleFollow = async () => {
+    if (!auth) return;
+
     if (myFollow) {
       await actions.deleteFollow({
         data: {
@@ -54,14 +66,104 @@ const ProfileTemplate = ({ user, auth, actions }: ProfileTemplatePropsType) => {
 
   const handlerProfileClick = () => {
     if (isMe) {
-      alert('profile edit modal');
+      setIsShowUpload(true);
     } else {
-      alert('large profile image');
+      // alert('large profile image');
+      setIsShowProfile(true);
     }
+  };
+
+  const handleOnSaveProfileImage = async (file: File) => {
+    // 동작
+    setIsLoadingUpload(true);
+
+    const form = new FormData();
+
+    form.append('isCover', 'false');
+    form.append('image', file);
+
+    await actions.uploadPhoto({
+      headers: {
+        'Content-Type': 'multipart/form-data',
+      },
+      data: form,
+    });
+
+    await actions.requestUser();
+
+    setIsShowUpload(false);
+    setIsLoadingUpload(false);
+  };
+
+  const handleOnCancelProfileImage = () => {
+    setIsShowUpload(false);
   };
 
   return (
     <>
+      <>
+        <Modal
+          visible={isShowUpload}
+          handleClose={() => setIsShowUpload(false)}
+        >
+          <ProfileImageUpload
+            currentImageUrl={user.image}
+            onSave={handleOnSaveProfileImage}
+            onCancel={handleOnCancelProfileImage}
+            disabled={isLoadingUpload}
+          />
+        </Modal>
+        <Modal
+          visible={isShowProfile}
+          handleClose={() => setIsShowProfile(false)}
+        >
+          <Modal.Header handleClose={() => setIsShowProfile(false)}>
+            <Text weight={600}>프로필 이미지</Text>
+          </Modal.Header>
+          <Image
+            width='100%'
+            height='50%'
+            src={user.image}
+            alt={`${user.fullName}'s cover image`}
+          />
+        </Modal>
+        <Modal
+          visible={isShowFollower}
+          handleClose={() => setIsShowFollower(false)}
+        >
+          <Modal.Header handleClose={() => setIsShowFollower(false)}>
+            팔로워
+          </Modal.Header>
+          <Modal.Body>
+            {user.followers.length === 0 && <Text>팔로워가 없습니다.</Text>}
+            {user.followers.map((follow) => (
+              <UserListItem
+                key={follow._id}
+                id={follow.user}
+              />
+            ))}
+          </Modal.Body>
+        </Modal>
+        <Modal
+          visible={isShowFollowing}
+          handleClose={() => setIsShowFollowing(false)}
+        >
+          <Modal.Header handleClose={() => setIsShowFollowing(false)}>
+            팔로잉
+          </Modal.Header>
+          <Modal.Body>
+            {user.following.length === 0 && (
+              <Text>팔로잉하고 있는 사용자가 없습니다.</Text>
+            )}
+            {user.following.map((follow) => (
+              <UserListItem
+                key={follow._id}
+                id={follow.user}
+              />
+            ))}
+          </Modal.Body>
+        </Modal>
+      </>
       <ProfileContainer maxWidth='md'>
         <Text
           size='3xl'
@@ -90,6 +192,7 @@ const ProfileTemplate = ({ user, auth, actions }: ProfileTemplatePropsType) => {
           <Flex
             direction='column'
             justifyContent='flex-end'
+            style={{ flexShrink: 1 }}
           >
             <Text
               size='2xl'
@@ -127,7 +230,8 @@ const ProfileTemplate = ({ user, auth, actions }: ProfileTemplatePropsType) => {
                   stroke={2.5}
                 />
               }
-              onClick={() => alert('my follower list')}
+              onClick={() => setIsShowFollower(true)}
+              disabled={!auth}
             >
               팔로워
             </Button>
@@ -150,6 +254,7 @@ const ProfileTemplate = ({ user, auth, actions }: ProfileTemplatePropsType) => {
                 )
               }
               onClick={handleFollow}
+              disabled={!auth}
             >
               {myFollow ? '팔로우 해제' : '팔로우'}
             </Button>
@@ -165,7 +270,7 @@ const ProfileTemplate = ({ user, auth, actions }: ProfileTemplatePropsType) => {
                   stroke={2.5}
                 />
               }
-              onClick={() => alert('my following list')}
+              onClick={() => setIsShowFollowing(true)}
             >
               팔로잉
             </Button>
@@ -188,8 +293,8 @@ const ProfileTemplate = ({ user, auth, actions }: ProfileTemplatePropsType) => {
         </Flex>
       </ProfileContainer>
       <PostContainer maxWidth='md'>
-        {user.posts.map((el) => (
-          <div>{el._id}</div>
+        {user.posts.map((post) => (
+          <div key={post._id}>{post._id}</div>
         ))}
       </PostContainer>
     </>
